% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(name = "World")
}
\arguments{
\item{name}{A character string; defaults to "World".}
}
\value{
Prints a greeting message.
}
\description{
Prints a greeting message.
}
